Update eligibility.js
Hey David, I've updated a few of the small discrepancies between this and the Axure prototype. This includes adding a hyperlink, but I'm not sure if you'd prefer this to be a local link, rather than an absolute link? 

I've also made a couple of other small wording changes, which I noticed when reviewing the content with fresh eyes. 

Hopefully that's all ok within the code!