Simplify featured errors logic
All errors in the current allow list have includeBaseError.
This means that for now we can make the assumption that all errors in
this allow-list will exclude the base error. Keen to avoid having a code
path that is not being executed.

Also I spotted that we can avoid the mutation and regular filtering
required by setting an isFeatured flag on the message by returning a
dedicated featuredMessages property from the validation function