Simplify joi-extensions
We had a code smell in our joi-extensions where we were relying on
parent state in order to validate related fields, making these
validations tightly coupled on the awards for all form.

This was based on the assumption that it was not possible to pass a
Joi.Ref as an argument to a validation, which **is** possible after all.

This PR reworks our extensions (with additional tests where needed to back up the
changes) to avoid any dependency on internal state objects