Enforce four digit year for org start date
Currently it is possible to enter a short year for the organisation
start date, such as 11 06. This is treated as 6AD which is a valid year
but not a useful one. Salesforce requires years to be four digits so
enforcing this here. Error message text needs reviewing and then
translating before this can be merged.