Fixup divergent use of relatedContent macro
A `usePromo` flag could be passed into the `relatedContent` macro making them behave very differently.

There were only two types of use:
1. homepage related news
2. flexible content related entries.

Noticed that the related-content classes were not even needed for the flexible content use-case. I can just be a flex-grid. So I've been able to simplify this to:

1. Inline the related-content markup into the homepage for that use-case
2. Simplify the flexible content use-case to a set of flex-grid classes
3. Delete the relatedContent macro