Consistent use of console logger during test runs
Sometimes seeing local tests runs try and fail to log to cloudwatch.

After doing some digging this is because the node envionrment is set to `test` which also determines which config file to use.
This means that cloudwatch logs are enabled as the app uses the same config as the test server.

The longer term fix for this would be to re-configure the staging site to used something other than test as the environment.
Basically: https://glebbahmutov.com/blog/do-not-use-node-env-for-staging/

In the short term if rejigged the logger config to include a check for an explicit `IS_TEST_RUN` variable to make sure the console logger is always used in these cases.