Update eligibility title
Needed for this title to make sense. Need to revisit this as the
eligibility checker can technically be reused, but we only use it for
this one form now and this is the most direct change to correct the copy
issue. Will revisit this to move the title into the eligibility model.