Disable latest funding programmes
Right now two of the three latest programmes are closed to applications
until further notice. Because this list is dynamic it's very misleading.
For now the simplest thing to do is to disable this list. If we
reinstate this we should consider making this a featured/manual list
rather than based on published date.