Update .compare() validation to compare normalised values
Spotted our `.compare()` validation is a little lax. Previous behaviour was to do a shallow equality check on the objects. This is fine if the object has exactly the same values e.g.

```js
isEqual({ a: 'one', b: 'two' }, { a: 'one', b: 'two' }) // true
```

But fails for cases where the keys contained extra spaces or a mix of cases:

```js
isEqual({ a: ' oNe ', b: 'twO' }, { a: 'one', b: 'two' }) // false
```

Updating this check to compare the list of normalised values instead. The trade-off here is that assumes all object keys can be stringified. For our use case this is a reasonable trade-off as we only use this check for comparing nested text fields like addresses and names.