Fix references needing to go one way
Unfortunately references in extensions can only go one way in the latest
version of joi. This was true previously for regular references but now
applied to extensions too.

This is the main design trade-off we are having to make in this upgrade.
Validation rules for same name and address now need to be one-way so
only associating the error with the second contact.